Added open source motion
Just added the open source motion 6f361693a7b248730b41d4292f89dc6f6f166bc8 give the dire situation we are in and the need for additional exchanges now we've lost BTER and Excoin. The risk of a copycat still exists but the risk of Nu's own existence are weighing in more.